Revert "lib: special handling of dongle disconnect for OS != win32"
This reverts commit f1803fe10f5017e5bf3528d69229d37472a024c8.

Commit f1803fe1 was an early attempt to fix libusb transfers in the
specific case of a disconnected dongle.  A new approach is more robust
(and handles other cancellation causes), but the reverted change also
introduced a behavioral difference in that non-windows hosts aborted
only on repeated failures of a specific type.

The restored behavior is that any error encountered during an
asynchronous transfer cancels the transfer.

Ref: http://lists.osmocom.org/pipermail/osmocom-sdr/2013-February/000481.html
Signed-off-by: Peter A. Bigot <pab@pabigot.com>